# strava_analytics/web/openapi.yaml
openapi: "3.1.0"
info:
  title: MERON API
  description: >
    Personal training data API. Provides fitness metrics, activity CRUD,
    training plans, race predictions, and Strava sync. All distances in
    miles, weights in lbs, elevations in feet. Times in seconds unless
    otherwise noted.
  version: "2.0.0"
servers:
  - url: http://localhost:8051
    description: Local standalone API server (strava-api)
  - url: http://localhost:8050
    description: Local Dash + API combined server (strava-web)
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/activities:
    get:
      tags: [activities]
      operationId: getRecentActivities
      summary: Recent activities
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 14 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RecentActivity" }
    post:
      tags: [activities]
      operationId: createActivity
      summary: Create a manual activity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActivityCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/activities/feed:
    get:
      tags: [activities]
      operationId: getActivityFeed
      summary: Cursor-paginated activity feed for the mobile infinite scroll
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
          description: Opaque cursor from a previous response. Omit for page 1.
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: One page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActivityFeedPage" }
  /api/activities/{activity_id}:
    parameters:
      - name: activity_id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [activities]
      operationId: getActivity
      summary: Fetch a single enriched activity
      responses:
        "200":
          description: Activity detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [activities]
      operationId: patchActivity
      summary: Update an activity (or add a manual override for Strava rows)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActivityPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  updated:
                    type: array
                    items: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [activities]
      operationId: deleteActivity
      summary: Soft-delete an activity
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  deleted: { type: boolean }
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/routes/{filename}:
    parameters:
      - name: filename
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [routes]
      operationId: getRouteDetail
      summary: Polyline + streams for a single activity
      responses:
        "200":
          description: Route
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouteDetail" }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Sign in with username + password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Session cookie set
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/SessionUser" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /oauth/strava/callback:
    get:
      tags: [oauth]
      operationId: oauthCallback
      summary: OAuth callback — exchanges code for tokens
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        "302":
          description: Redirect back to /settings#strava-connected
        "400":
          description: State mismatch or missing code
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Read key for GET; write key for POST/PATCH/DELETE.
  responses:
    BadRequest:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
          required: [code, message]
...
    RecentActivity:
      type: object
      properties:
        date: { type: string, format: date }
        type: { type: string }
        name: { type: string }
        distance_mi: { type: number }
        pace: { type: string }
        pace_decimal: { type: number }
        avg_hr: { type: number }
        duration_s: { type: number }
        elevation_ft: { type: number }
        run_type: { type: string }
    FeedItem:
      type: object
      properties:
        _id: { type: integer }
        date: { type: string, format: date }
        type: { type: string }
        name: { type: string }
        distance_mi: { type: number }
        pace: { type: string }
        pace_decimal: { type: number }
        avg_hr: { type: number }
        duration_s: { type: number }
        elevation_ft: { type: number }
        run_type: { type: string }
    ActivityFeedPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FeedItem" }
        next_cursor:
          type: string
          nullable: true
          description: Pass to the next GET /activities/feed; null means end of feed.
    Activity:
      type: object
      description: Full enriched activity record. Fields vary by activity type.
      additionalProperties: true
      properties:
        _id: { type: integer }
        date: { type: string, format: date }
        type: { type: string }
        name: { type: string }
        distance_m: { type: number }
        distance_mi: { type: number }
        moving_time_s: { type: number }
        elapsed_time_s: { type: number }
        avg_hr: { type: number }
        max_hr: { type: number }
        avg_watts: { type: number }
        calories: { type: number }
        elevation_gain_m: { type: number }
    ActivityCreate:
      type: object
      required: [type, start_time]
      properties:
        type: { type: string }
        start_time: { type: string, format: date-time }
        name: { type: string }
        description: { type: string }
        gear: { type: string }
        elapsed_time_s: { type: number, minimum: 0 }
        moving_time_s: { type: number, minimum: 0 }
        distance_m: { type: number, minimum: 0 }
        elevation_gain_m: { type: number }
        max_hr: { type: number }
        avg_hr: { type: number }
        avg_watts: { type: number, minimum: 0 }
        calories: { type: number, minimum: 0 }
        weather_condition: { type: string }
        weather_temp_c: { type: number }
    ActivityPatch:
      type: object
      description: Partial update — every field optional.
      additionalProperties: false
      properties:
        name: { type: string }
        type: { type: string }
        description: { type: string }
        gear: { type: string }
        filename: { type: string }
        start_time: { type: string, format: date-time }
        elapsed_time_s: { type: number }
        moving_time_s: { type: number }
        distance_m: { type: number }
        max_speed_ms: { type: number }
        avg_speed_ms: { type: number }
        elevation_gain_m: { type: number }
        elevation_loss_m: { type: number }
        elevation_low_m: { type: number }
        elevation_high_m: { type: number }
        max_hr: { type: number }
        avg_hr: { type: number }
        avg_watts: { type: number }
        calories: { type: number }
        relative_effort: { type: number }
        grade_adj_distance_m: { type: number }
        total_steps: { type: number }
        training_load: { type: number }
        intensity: { type: number }
        weather_condition: { type: string }
        weather_temp_c: { type: number }
        competition: { type: string }
        strava_with_kid: { type: integer }
...
    SessionUser:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        display_name: { type: string, nullable: true }
        is_admin: { type: boolean }
...
    RouteDetail:
      type: object
      properties:
        filename: { type: string }
        polyline:
          type: array
          items:
            type: array
            items: { type: number }
        streams:
          type: object
          properties:
            timestamps:
              type: array
              items: { type: string }
            heart_rate:
              type: array
              items: { type: number, nullable: true }
            altitude_m:
              type: array
              items: { type: number, nullable: true }
            speed_m_per_s:
              type: array
              items: { type: number, nullable: true }
...